Added focus line. All non-slider tooltip elements have been added.
