+++ assets/src/styles/components/dv-hydrograph.scss
@@ -107,12 +107,10 @@ svg {
     fill: black;
     opacity: .6;
   }
-  .focus {
-    line {
-      opacity: .5;
-      stroke: #436baf;
-      stroke-width: 1px;
-      stroke-dasharray: 5, 5;
-    }
+  .focus-line {
+    opacity: .5;
+    stroke: black;
+    stroke-width: 1px;
+    stroke-dasharray: 5, 5;
   }
 }
